:memo: Add documentation
* Add CONTRIBUTING.md
* :page_facing_up: ADD LICENSE
* Update README
* :fire: Remove SoftBreak

* :heavy_plus_sign: Add docz for documentation writing
