+++ .gitignore
@@ -3,6 +3,9 @@
 # dependencies
 /node_modules
 
+# docs
+/.docz
+
 # testing
 /coverage
