+++ README.md
@@ -1,9 +1,15 @@
-# Edtr.io @next
-Major revamp of the current Edtr.io, based on [Slate.js](https://github.com/ianstormtaylor/slate).
+# Edtr.io  ![Version](https://img.shields.io/badge/version-1.0.0-black.svg) ![[Gitmoji](https://gitmoji.carloscuesta.me)](https://img.shields.io/badge/gitmoji-%20😜%20😍-FFDD67.svg?style=flat-square) 
 
-> Current status: **Work In Progress**
+A rich text editor for easily creating content in schools based on [Slate.js](https://github.com/ianstormtaylor/slate).
 
-## Current features
+## Getting started
+```shell
+yarn install
+yarn start
+# go to localhost:3000 if it doesnt open automatically
+```
+
+## Features
 - Hovering menu (**bold**, *italic*, <u>underline</u>, `code`, h1, h2)
 - Markdown shortcuts (h1-h5 `#`, ul `-`, blockquotes `>`)
 - Auto-detecting urls (need `http(s)://` prefix atm)
@@ -12,15 +18,14 @@ Major revamp of the current Edtr.io, based on [Slate.js](https://github.com/ians
 - Button to add Geogebra applets
 - Video integration (paste youtube video link and it'll be auto-converted)
 
-## Setup
-```shell
-yarn install
-yarn start
-# go to localhost:3000 if it doesnt open automatically
-```
+## Documentation
+> TODO: add link to docs
+The docs can be found [here](#).
 
+## Contribute
+We very much welcome any kinds of contributions! Have a look at our [issue tracker](https://github.com/schul-cloud/edtrio/issues) or [project board roadmap](https://github.com/schul-cloud/edtrio/projects/1) for inspiration.
+- Issue Tracker: https://github.com/schul-cloud/edtrio/issues
+- Source Code: https://github.com/schul-cloud/edtrio
 
----
-This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
-
-You can find the some information on how to perform common tasks [here](https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md).
+## License
+The project is licensed under the [MIT License](LICENSE).
