+++ docs/guides/overview.mdx
@@ -1,0 +1,11 @@
+---
+name: Overview
+route: /guides
+menu: 2. Guides
+order: 20
+---
+
+# Guides Overview
+Hm don't really know what to put here.
+
+Head on over to learn [How to create a plugin](/guides/create-plugin)!
