+++ docs/index.mdx
@@ -1,0 +1,22 @@
+---
+name: Overview
+route: '/'
+---
+
+# Overview
+
+[Edtr.io](https://github.com/schul-cloud/edtrio) is a rich text editor for easily creating content in schools based on [Slate.js](https://github.com/ianstormtaylor/slate).
+
+## Getting started
+```shell
+yarn install
+yarn start
+# go to localhost:3000 if it doesnt open automatically
+```
+
+## Development
+> ⚠️ Before starting to work on Edtr.io it is **highly recommended** to go trough the [Slate.js Walktroughs and Guides](https://docs.slatejs.org/)
+
+You can learn more about the general structure of Edtr.io in section [1. Structure](/structure). There are a multitude of guides on how to work with the editor in section [2. Guides](/guides). Finally, there is a summary of all plugins in section [3. Plugins](/plugins).
+
+Before starting development please also have a look at our [Contributing Guidelines](https://github.com/schul-cloud/edtrio/blob/master/CONTRIBUTING.md) - it's very short, promise! ❤️
