+++ docs/plugins/overview.mdx
@@ -1,0 +1,26 @@
+---
+name: Overview
+route: /plugins/
+menu: 3. Plugins
+order: 200
+---
+
+# Plugins Overview
+These are all custom developed plugins currently included in Edtr.io:
+
+| Plugin               | Description                                  |
+| -------------------- | -------------------------------------------- |
+| `add-section`        | Exports button component to append a section |
+| `text-menu`          | Adds popup text menu (bold, italic, ...)     |
+| `plus-menu`          | Adds small plus to row to add interactive content |
+| `code-block`         | Adds simple code blocks to the editor        |
+| `geogebra`           | Adds geogebra to the editor                  |
+| `iframe`             | Adds iframes to the editor                   |
+| `image`              | Adds images to the editor                    |
+| `title`              | Adds a title to the editor                   |
+| `section`            | Adds sections and p tags to the editor       |
+| `url-handler`        | Automatically tries to inline urls content (e.g. YouTube videos) |
+| `auto-url`           | Automatically detects urls and highlights them |
+| `markdown-shortcuts` | Automatically turns markdown into rich elements |
+| `helpers`            | Set of helper functions (e.g. `insertParagraph()`) |
+| `download-file`      | Adds up/download functionality for files (UNUSED ATM) |
