+++ docs/structure/document.mdx
@@ -1,8 +1,18 @@
-# Schema
-The schema of edtr.io docs is closely aligned to the HTML5 standard to not introduce further complexity.
+---
+name: Document Schema
+route: /structure/document
+menu: 1. Structure
+order: 100
+---
 
-# Nodes
-## Types
+# Document Schema
+The schema of edtr.io documents is closely aligned to the HTML5 standard to not introduce further complexity.
+
+## DocumentViewer
+To have a look at the underlying document during development, a [`DocumentViewer`](https://github.com/schul-cloud/edtrio/tree/master/src/dev-document-viewer) is included and automatically displayed at the bottom of the page.
+
+## Nodes
+### Types
 - `title`: Title of the document (`min: 1, max: 1`)
 - `section`: Thematic grouping of content, typically with a heading
 - `code`: Piece of computer code
@@ -16,7 +26,7 @@ The schema of edtr.io docs is closely aligned to the HTML5 standard to not intro
 - `video`: Video embeds (e.g. YouTube)
 - `geogebra`
 
-## Structure
+### Structure
 There are two higher level node types: `title` and `section`.
 Each document must have exactly one `title` node. It may only contain a `text` node.\
 It is followed by 1:n many `section` nodes. `section` nodes can contain any type of node.
@@ -31,9 +41,9 @@ It is followed by 1:n many `section` nodes. `section` nodes can contain any type
 ```
 
 
-# Mark types
+## Mark types
 - `strong`
 - `em`
 - `cite`
 
-For a more detailed documentation also take a look at `schema.js`.
+For a more detailed documentation also [take a look at `schema.js`](https://github.com/schul-cloud/edtrio/blob/master/src/schema.js).
