+++ doczrc.js
@@ -1,0 +1,9 @@
+export default {
+    title: 'Edtr.io Documentation',
+    description: '',
+    themeConfig: {
+        colors: {
+            primary: 'tomato',
+        },
+    }
+}
